minor review suggestions adjustments
